resolved runtime <-> loader dependency, path conversion moved to runtime. extended configuration loading to support sections in configuration files. prepared for introduction of POSIX building with cl.exe
git-svn-id: https://svn.code.sf.net/p/parity/code/trunk@108 26386e8f-9d3d-0410-aaca-840ac0c00a30

+++ parity.binary/CoffFileHeader.cpp
@@ -353,7 +353,7 @@ namespace parity
 				//
 				for(Symbol::IndexedSymbolMap::iterator sym = map_.begin(); sym != map_.end(); ++sym)
 				{
-					if(sym->second.getSectionNumber() == it->first
+					if(sym->second.getSectionNumber() == (short)it->first
 						&& sym->second.getStorageClass() == Symbol::ClassStatic
 						&& sym->second.getName() == it->second.getName())
 					{
